+++ psx.ps1
@@ -1,4 +1,4 @@
-#requires -version 5.1
+﻿#requires -version 5.1
 $toolDir = Split-Path -Parent $MyInvocation.MyCommand.Path
-& (Join-Path $toolDir 'psctx.ps1') @args
+& (Join-Path $toolDir 'PSX.Core.ps1') @args
 exit $LASTEXITCODE